pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "DENDENICH",email = "daniil.gol26@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "docutils (>=0.22.3,<0.23.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
